modified for numpy

+++ source/baselines/baseline_ridi.py
@@ -72,7 +72,7 @@ class CvSVR(SVRBase):
                      target.astype(np.float32))
 
     def predict(self, feature):
-        return self.m.predict(feature.astype(np.float32))[1].astype(np.float).ravel()
+        return self.m.predict(feature.astype(np.float32))[1].astype(np.float64).ravel()
 
     def save(self, out_dir):
         self.m.save(osp.join(out_dir, 'regressor_{}.yaml'.format(self.chn)))
@@ -219,7 +219,7 @@ def grid_search(args):
 
 
 def recon_traj_with_preds(dataset, preds, seq_id=0, **kwargs):
-    ind = np.array([i[1] for i in dataset.index_map if i[0] == seq_id], dtype=np.int)
+    ind = np.array([i[1] for i in dataset.index_map if i[0] == seq_id], dtype=np.int32)
     ts = dataset.ts[seq_id]
     dt = np.mean(ts[ind[1:]] - ts[ind[:-1]])
